fix: Split eval_test.exs into multiple focused test files
Splits the large eval_test.exs (1078 lines, 114 tests) into 6 focused test files:
- eval_data_types_test.exs: literal, vector, map, set evaluation
- eval_control_flow_test.exs: variable access, and/or, if, let, where, predicates
- eval_functions_test.exs: fn, closures, destructuring, multi-arity, builtins
- eval_errors_test.exs: error propagation in structures, bindings, calls
- eval_memory_test.exs: memory threading through evaluation
- eval_sets_test.exs: set predicates, constructor, collection operations

Follows the pattern from issue #95 for organizing test files by functional area.
Each file is self-contained with proper module setup and helper functions.
Test count remains at 114 tests.

Fixes #173
